backend/kesaseteli/requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
apipkg==1.5
    # via execnet
appdirs==1.4.4
    # via black
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via ipython
black==21.5b2
    # via -r requirements-dev.in
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
click==8.0.1
    # via black
coverage==5.5
    # via pytest-cov
decorator==5.1.1
    # via ipython
execnet==1.8.1
    # via pytest-xdist
flake8==3.9.2
    # via -r requirements-dev.in
freezegun==1.1.0
    # via pytest-freezegun
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via traitlets
ipython==7.24.1
    # via -r requirements-dev.in
isort==5.8.0
    # via -r requirements-dev.in
jedi==0.18.0
    # via ipython
langdetect==1.0.9
    # via -r requirements-dev.in
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.18
    # via ipython
psutil==5.8.0
    # via pytest-xdist
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest-cov==2.12.1
    # via -r requirements-dev.in
pytest-django==4.4.0
    # via -r requirements-dev.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-freezegun==0.4.2
    # via -r requirements-dev.in
pytest-xdist[psutil]==2.2.1
    # via -r requirements-dev.in
pytest==6.2.4
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-freezegun
    #   pytest-xdist
python-dateutil==2.8.1
    # via freezegun
regex==2021.4.4
    # via black
requests-mock==1.9.3
    # via -r requirements-dev.in
requests==2.25.1
    # via requests-mock
six==1.16.0
    # via
    #   langdetect
    #   python-dateutil
    #   requests-mock
toml==0.10.2
    # via
    #   black
    #   pytest
    #   pytest-cov
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.5
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.1
    # via -r requirements-dev.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
